+++ patzilla-ui/navigator/components/nataraja/index.ts
@@ -1,0 +1,3 @@
+/* eslint import/no-extraneous-dependencies: 0 */
+import Nataraja from './nataraja';
+export { Nataraja };